+++ src/wl_tests/test_ngram.py
@@ -33,11 +33,11 @@ def test_ngram():
     # Exhaust all n-grams
     main.settings_custom['ngram']['search_settings']['search_settings'] = False
 
-    for i, file_test in enumerate(main.settings_custom['files']['files_open']):
-        for file in main.settings_custom['files']['files_open']:
+    for i, file_test in enumerate(main.settings_custom['file_area']['files_open']):
+        for file in main.settings_custom['file_area']['files_open']:
             file['selected'] = False
 
-        main.settings_custom['files']['files_open'][i]['selected'] = True
+        main.settings_custom['file_area']['files_open'][i]['selected'] = True
 
         print(f'''Testing file "{file_test['name']}"... ''', end = '')